alx_travel_app/listings/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import ListingViewSet, BookingViewSet, PaymentViewSet, sample_api, chapa_webhook

router = DefaultRouter()
router.register(r'listings', ListingViewSet)
router.register(r'bookings', BookingViewSet)
router.register(r'payments', PaymentViewSet)

urlpatterns = [
    path('', include(router.urls)),
    path('sample/', sample_api, name='sample-api'),
    path('webhook/chapa/', chapa_webhook, name='chapa-webhook'),
]
