alx_travel_app/requirement.txt
Django
djangorestframework
django-cors-headers
drf-yasg
django-environ
celery
psycopg2-binary
requests
pyngrok